+++ Library/CloudVOffice.Core/Domain/HR/Master/Department.cs
@@ -12,9 +12,9 @@ namespace CloudVOffice.Core.Domain.HR.Master
 		public string DepartmentName { get; set; }
 		public int? Parent { get; set; }
 		public bool IsGroup { get; set; }
-		public int CreatedBy { get; set; }
+		public Int64 CreatedBy { get; set; }
 		public DateTime CreatedDate { get; set; }
-		public int? UpdatedBy { get; set; }
+		public Int64? UpdatedBy { get; set; }
 		public DateTime? UpdatedDate { get; set; }
 		public bool Deleted { get; set; }
 		public List<Department> Children { get; set; }
