+++ Library/CloudVOffice.Core/Domain/HR/Master/Designation.cs
@@ -11,9 +11,9 @@ namespace CloudVOffice.Core.Domain.HR.Master
 		public int DesignationId { get; set; }
 		public string DesignationName { get; set; }
 		public string Description { get; set; }	
-		public int CreatedBy { get; set; }
+		public Int64 CreatedBy { get; set; }
 		public DateTime CreatedDate { get; set; }
-		public int? UpdatedBy { get; set; }
+		public Int64? UpdatedBy { get; set; }
 		public DateTime? UpdatedDate { get; set; }
 		public bool Deleted { get; set; }
 	}
