+++ Library/CloudVOffice.Core/Domain/Pemission/Application.cs
@@ -15,9 +15,9 @@ namespace CloudVOffice.Core.Domain.Pemission
         public int? Parent { get; set; }
         public bool IsGroup { get; set; }
         public string? Url { get; set; }
-        public int CreatedBy { get; set; }
+        public Int64 CreatedBy { get; set; }
         public DateTime CreatedDate { get; set; }
-        public int? UpdatedBy { get; set; }
+        public Int64? UpdatedBy { get; set; }
         public DateTime? UpdatedDate { get; set; }
         public bool Deleted { get; set; }
