+++ Library/CloudVOffice.Core/Domain/Users/Role.cs
@@ -12,9 +12,9 @@ namespace CloudVOffice.Core.Domain.Users
         public int RoleId { get; set; }
         public string RoleName { get; set; }
        
-		public int CreatedBy { get; set; }
+		public Int64 CreatedBy { get; set; }
 		public DateTime CreatedDate { get; set; }
-		public int? UpdatedBy { get; set; }
+		public Int64? UpdatedBy { get; set; }
 		public DateTime? UpdatedDate { get; set; }
 		public bool Deleted { get; set; }
 		public ICollection<UserRoleMapping> UserRoleMappings { get; set; }
